add[BE]: Add DmController
