+++ backend/src/api/channels/channels.module.ts
@@ -4,10 +4,11 @@ import { ChannelsService } from './channels.service';
 import { ChannelsController } from './channels.controller';
 import { PrismaService } from '../../prisma/prisma.service';
 import { AuthModule } from '../auth/auth.module';
+import { DmController } from './dm.controller';
 
 @Module({
   imports: [AuthModule],
   providers: [ChannelsGatway, ChannelsService, PrismaService],
-  controllers: [ChannelsController],
+  controllers: [ChannelsController, DmController],
 })
 export class ChannelsModule {}
